hw8 ready

+++ src/main/java/az/edu/turing/happyfamilyv1/Human.java
@@ -5,7 +5,9 @@ import java.time.Instant;
 import java.time.LocalDate;
 import java.time.Period;
 import java.time.ZoneId;
-import java.util.*;
+import java.util.Date;
+import java.util.HashMap;
+import java.util.Map;
 
 public class Human {
 
@@ -53,8 +55,8 @@ public class Human {
         this.surname = surname;
     }
 
-    public long getBirthDate() {
-        return birthDate;
+    public Integer getBirthDate() {
+        return Math.toIntExact(birthDate);
     }
 
     public void setBirthYear(long birthDate) {
